Update Ghost to Trusty, cleanup

+++ services/ghost/start.sh
@@ -2,21 +2,11 @@
 
 set -e
 
-mkdir -p /data
+[ ! -f /data/config.js ] && cp /ghost/config.example.js /data/config.js
+[ ! -d /content.default ] && mv /ghost/content /content.default
+[ ! -d /data/content ] && cp -r /content.default /data/content
 
-if [ ! -f /data/config.js ]; then
-  cp /ghost/config.example.js /data/config.js
-fi
+[ ! -L /ghost/content ] && ln -sf /data/content /ghost/content
+[ ! -L /ghost/config.js ] && ln -sf /data/config.js /ghost/config.js
 
-if [ ! -d /content.default ]; then
-  mv /ghost/content /content.default
-fi
-
-if [ ! -d /data/content ]; then
-  cp -r /content.default /data/content
-fi
-
-ln -sf /data/content /ghost/content
-ln -sf /data/config.js /ghost/config.js
-
-cd /ghost && npm start --production
+npm start --production
